stabilize test_tus_crash_native
